requirements.txt
delta-spark==3.2.0
esdbclient==1.0.25
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
importlib_metadata==7.1.0
numpy==1.26.4
pandas==2.2.2
protobuf==4.25.3
py4j==0.10.9.7
pyarrow==16.0.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
zipp==3.18.1
